feat(engine): Z-Image CUDA path and Linux CUDA server release tar.gz
Enable Z-Image text-to-image on CudaContext (param collection, torch weight
assign, mflux-style CFG, on-device modules) and add PyInstaller packaging plus
GitHub Actions job to build danqing-studio-linux-cuda-x86_64-*.tar.gz alongside
macOS desktop releases. Includes Flux.1 pipeline/registry work and benchmark
updates for CUDA-capable models.

+++ backend/engine/families/z_image/text_encoder_cuda.py
@@ -7,8 +7,8 @@ from typing import Any
 def zimage_prepare_torch_ids(ctx: Any, input_ids_np, attention_mask_np):
     import torch
 
-    input_ids = torch.tensor(input_ids_np, dtype=torch.int32, device=ctx._device)
-    attention_mask = torch.tensor(attention_mask_np, dtype=torch.float32, device=ctx._device)
+    input_ids = torch.tensor(input_ids_np, dtype=torch.long, device=ctx._device)
+    attention_mask = torch.tensor(attention_mask_np, dtype=torch.long, device=ctx._device)
     return input_ids, attention_mask
